requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.1
streamlit==1.28.0
plotly==5.17.0
shap==0.42.1
evidently==0.4.12
mlflow==2.10.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
matplotlib==3.10.1
xgboost==1.8.0
pytest==6.2.5
pyyaml==5.1.2